Made changes, sorry about the wait
Made changes to section 5. It's kind of brief because these have no real impact on our game or are related to our game. 
For instance, in 5.1: We were never given a time complexity/space complexity so there really isn't a concern for that. The game just needs to run and not crash
5.2: There's no safety related problems (other than playing too long, I guess?)
5.3: There are no security requirements because there's no database/user info/private info
5.4: Software quality, we weren't necessarily told about how well the game needs to look, it just needs to look like chess and work
5.5: We have no business rules, this game isn't going public

Let me know if I should add anything